+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 
 - Fix error when VPC and EC2 instances are created together ([#13](https://github.com/WildBeavers/terraform-aws-ec2-instance/issues/13)).
   Parameter `subnet_id` is deprecated and can not be used together with `subnet_ids`.
+- when variables has existed for a single value and a list of values
+  then the single value variable has been removed (`private_ip`, `subnet_id`)  
+  => variable `subnet_ids` is now required
+- deprecated outputs have been removed
+- instances are returned as a list of instances
+- Route53 A records are added to the outputs
+- use a map for volumes and volumes attachments instead of a list  
+  migrate existing installations with `tf state mv`:
+  
+      # replace ${index}, ${hostname} and ${volumename} accordingly
+      tf state mv 'module.ec2.aws_ebs_volume.this[${index}]' 'module.ec2.aws_ebs_volume.this["${hostname}.${volumename}"]'
 
 <a name="v2.7.0"></a>
 ## [v2.7.0] - 2019-09-06
@@ -13,6 +24,8 @@
 - Support new Terraform 0.12 feature: EC2 instance as output  ([#4](https://github.com/WildBeavers/terraform-aws-ec2-instance/issues/4))
 - The format string for generating the hostname of an instance can be overridden ([#9](https://github.com/WildBeavers/terraform-aws-ec2-instance/issues/9))
 - support suffix for volume tag `Name` ([#5](https://github.com/WildBeavers/terraform-aws-ec2-instance/issues/5))
+- removed parameter `volume_tag_name_suffix` of variable `attached_block_device`
+- add required parameter `volume_name` to variable `attached_block_device`
 
 <a name="v2.6.0"></a>
 ## [v2.6.0] - 2019-07-21